requirements.txt
streamlit
torch
torchvision
numpy
pandas
Pillow
matplotlib
scipy
